azure/python/requirements.txt
appnope==0.1.2
autopep8==1.6.0
azure-core==1.23.0
azure-storage-blob==12.10.0
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.0
docopt==0.6.2
dotenv-python==0.0.1
entrypoints==0.3
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
isodate==0.6.1
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
msrest==0.6.21
nest-asyncio==1.5.4
numpy==1.22.2
oauthlib==3.2.0
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pyairtable==1.1.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.10.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
yarg==0.1.9
